requirements.txt
flask>=3.0
yfinance>=0.2
pandas>=2.0
matplotlib>=3.8
seaborn>=0.13
requests>=2.31
google-generativeai>=0.7
python-dotenv>=1.0
boto3>=1.34
